+++ dflib/src/test/java/org/dflib/RowSet_Select_RowsTailTest.java
@@ -47,7 +47,21 @@ public class RowSet_Select_RowsTailTest {
     }
 
     @Test
-    public void negative() {
+    public void negative1() {
+        DataFrame df = DataFrame.foldByRow("a", "b").of(
+                        1, "x",
+                        2, "y",
+                        3, "z")
+                .rowsTail(-1).select();
+
+        new DataFrameAsserts(df, "a", "b")
+                .expectHeight(2)
+                .expectRow(0, 1, "x")
+                .expectRow(1, 2, "y");
+    }
+
+    @Test
+    public void negative2() {
         DataFrame df = DataFrame.foldByRow("a", "b").of(
                         1, "x",
                         2, "y",